+++ test/Test.hs
@@ -62,7 +62,7 @@ modifyTestTime modtime = modify (\te -> te { teCurrentTime = modtime . teCurrent
 runTestEnv :: ConnectionSourceM (LogT IO) -> Logger -> TestEnv a -> IO a
 runTestEnv connSource logger m =
     (runLogT "consumers-test" logger)
-  . (runDBT connSource {- transactionSettings -} def)
+  . (runDBT connSource defaultTransactionSettings)
   . (\m' -> fst <$> (runStateT m' $ TestEnvSt $ UTCTime (ModifiedJulianDay 0) 0))
   . unTestEnv
   $ m
@@ -77,7 +77,8 @@ test = do
         []     -> defaultConnString
         (cs:_) -> cs
 
-  let connSettings                = def { csConnInfo = T.pack connString }
+  let connSettings                = defaultConnectionSettings
+                                    { csConnInfo = T.pack connString }
       ConnectionSource connSource = simpleSource connSettings
 
   withSimpleStdOutLogger $ \logger ->
@@ -120,14 +121,17 @@ test = do
 
       createTables :: TestEnv ()
       createTables = do
-        migrateDatabase {- options -} def {- extensions -} [] {- domains -} []
+        migrateDatabase defaultExtrasOptions
+          {- extensions -} [] {- composites -} [] {- domains -} []
           tables migrations
-        checkDatabase {- options -} def {- domains -} [] tables
+        checkDatabase defaultExtrasOptions
+          {- composites -} [] {- domains -} []
+          tables
 
       dropTables :: TestEnv ()
       dropTables = do
-        migrateDatabase {- options -} def {- extensions -} [] {- domains -} []
-          {- tables -} []
+        migrateDatabase defaultExtrasOptions
+          {- extensions -} [] {- composites -} [] {- domains -} [] {- tables -} []
           [ dropTableMigration jobsTable
           , dropTableMigration consumersTable ]